requirements.txt
certifi==2021.5.30
cvxopt==1.2.6
cvxpy==1.1.15
dill==0.3.4
ecos==2.0.7.post1
multiprocess==0.70.12.2
numpy==1.19.5
osqp==0.6.2.post0
python-constraint==1.4.0
qdldl==0.1.5.post0
scipy==1.5.4
scs==2.1.4
simanneal==0.5.0
six==1.16.0
